requirements.txt
watchdog
Pillow
progress